# docs/swagger.yaml
basePath: /api
definitions:
  models.PaginatedMeta:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  models.QurbanAnimalType:
    enum:
    - cow
    - goat
    type: string
    x-enum-varnames:
    - Cow
    - Goat
  models.QurbanOfferingListWithPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.QurbanOfferingResponse'
        type: array
      meta:
        $ref: '#/definitions/models.PaginatedMeta'
    type: object
  models.QurbanOfferingMutationResponse:
    properties:
      data:
        $ref: '#/definitions/models.QurbanOfferingResponse'
      message:
        example: Qurban Offering created successfully
        type: string
    type: object
  models.QurbanOfferingRequest:
    properties:
      animal_type:
        allOf:
        - $ref: '#/definitions/models.QurbanAnimalType'
        enum:
        - cow
        - goat
      capacity:
        type: integer
      name:
        maxLength: 100
        type: string
      price:
        type: number
      qurban_period_id:
        type: integer
      scheme_type:
        allOf:
        - $ref: '#/definitions/models.QurbanSchemeType'
        enum:
        - group
        - individual
      status:
        allOf:
        - $ref: '#/definitions/models.QurbanStatus'
        enum:
        - open
        - closed
    required:
    - animal_type
    - capacity
    - name
    - price
    - qurban_period_id
    - scheme_type
    type: object
  models.QurbanOfferingResponse:
    properties:
      animal_type:
        $ref: '#/definitions/models.QurbanAnimalType'
      capacity:
        type: integer
      confirmed_slots:
        type: integer
      created_at:
        type: string
      filled_slots:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      qurban_period_id:
        type: integer
      scheme_type:
        $ref: '#/definitions/models.QurbanSchemeType'
      status:
        $ref: '#/definitions/models.QurbanStatus'
      updated_at:
        type: string
    type: object
  models.QurbanPeriodListWithPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.QurbanPeriodResponse'
        type: array
      meta:
        $ref: '#/definitions/models.PaginatedMeta'
    type: object
  models.QurbanPeriodMutationResponse:
    properties:
      data:
        $ref: '#/definitions/models.QurbanPeriodResponse'
      message:
        example: Qurban Period created successfully
        type: string
    type: object
  models.QurbanPeriodOptionsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.QurbanPeriodResponse'
        type: array
    type: object
  models.QurbanPeriodRequest:
    properties:
      description:
        type: string
      end_date:
        type: string
      start_date:
        type: string
      year:
        type: integer
    required:
    - end_date
    - start_date
    - year
    type: object
  models.QurbanPeriodResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  models.QurbanSchemeType:
    enum:
    - group
    - individual
    type: string
    x-enum-varnames:
    - Group
    - Individual
  models.QurbanStatus:
    enum:
    - open
    - closed
    type: string
    x-enum-varnames:
    - Open
    - Closed
  models.SwaggerError:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: Qurban period not found
        type: string
    type: object
  models.SwaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.SwaggerError'
    type: object
  models.SwaggerValidationDetail:
    additionalProperties:
      type: string
    type: object
  models.SwaggerValidationError:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details:
        $ref: '#/definitions/models.SwaggerValidationDetail'
      message:
        example: Validation failed
        type: string
    type: object
  models.SwaggerValidationErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.SwaggerValidationError'
    type: object
host: localhost:4000
info:
  contact:
    email: dev@kurbankan.id
    name: Kurbankan Dev
  description: API server for Kurbankan qurban management platform.
  title: Kurbankan API
  version: "1.0"
paths:
  /api/migrate:
    post:
      consumes:
      - application/json
      description: Automatically migrate all database tables based on GORM models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Run database migration
      tags:
      - Migration
  /api/migrate/drop:
    delete:
      consumes:
      - application/json
      description: DROP ALL TABLES - USE WITH EXTREME CAUTION! This will delete all
        data.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Drop all database tables
      tags:
      - Migration
  /api/migrate/status:
    get:
      consumes:
      - application/json
      description: Check if all required database tables exist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      summary: Check migration status
      tags:
      - Migration
  /api/seed/clear:
    delete:
      consumes:
      - application/json
      description: Delete all data from provinces, regencies, districts, and villages
        tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear all seeded data
      tags:
      - Seeder
  /api/seed/run:
    post:
      consumes:
      - application/json
      description: Execute all seed SQL files in the correct order
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Run all seeders
      tags:
      - Seeder
  /api/seed/run/{seed}:
    post:
      consumes:
      - application/json
      description: Execute a specific seed SQL file
      parameters:
      - description: Seed name (provinces, regencies, districts, villages_1, villages_2)
        in: path
        name: seed
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Run specific seeder
      tags:
      - Seeder
  /api/seed/status:
    get:
      consumes:
      - application/json
      description: Check which tables have been seeded and data counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      summary: Check seeder status
      tags:
      - Seeder
  /qurban-offerings:
    get:
      description: Returns a paginated list of qurban offerings. Optionally filter
        by animal type and scheme type.
      parameters:
      - description: Filter by animal type (e.g. cow, goat)
        in: query
        name: animal_type
        type: string
      - description: Filter by scheme type (e.g. regular, premium)
        in: query
        name: scheme_type
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanOfferingListWithPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List qurban offerings
      tags:
      - Qurban Offerings
    post:
      consumes:
      - application/json
      description: Creates a new qurban offering for the authenticated mosque.
      parameters:
      - description: Qurban offering payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.QurbanOfferingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.QurbanOfferingMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SwaggerValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a qurban offering
      tags:
      - Qurban Offerings
  /qurban-offerings/{id}:
    delete:
      description: Deletes a qurban offering by ID for the authenticated mosque.
      parameters:
      - description: Qurban Offering ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanOfferingMutationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a qurban offering
      tags:
      - Qurban Offerings
    patch:
      consumes:
      - application/json
      description: Updates an existing qurban offering by ID for the authenticated
        mosque.
      parameters:
      - description: Qurban Offering ID
        in: path
        name: id
        required: true
        type: integer
      - description: Qurban offering payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.QurbanOfferingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanOfferingMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SwaggerValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a qurban offering
      tags:
      - Qurban Offerings
  /qurban-periods:
    get:
      description: Returns a paginated list of qurban periods. Optionally filter by
        year.
      parameters:
      - description: Filter by year (e.g. 2025)
        in: query
        name: year
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanPeriodListWithPaginationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List qurban periods
      tags:
      - Qurban Periods
    post:
      consumes:
      - application/json
      description: Creates a new qurban period for the authenticated mosque.
      parameters:
      - description: Qurban period payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.QurbanPeriodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.QurbanPeriodMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SwaggerValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a qurban period
      tags:
      - Qurban Periods
  /qurban-periods/{id}:
    delete:
      description: Deletes a qurban period by ID for the authenticated mosque.
      parameters:
      - description: Qurban Period ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanPeriodMutationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a qurban period
      tags:
      - Qurban Periods
    patch:
      consumes:
      - application/json
      description: Updates an existing qurban period by ID for the authenticated mosque.
      parameters:
      - description: Qurban Period ID
        in: path
        name: id
        required: true
        type: integer
      - description: Qurban period payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.QurbanPeriodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanPeriodMutationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SwaggerValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a qurban period
      tags:
      - Qurban Periods
  /qurban-periods/options:
    get:
      description: Returns a list of qurban period options for dropdowns or selection
        inputs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QurbanPeriodOptionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List qurban period options
      tags:
      - Qurban Periods
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"